+++ Procfile.dev
@@ -1,9 +1,9 @@
 # Run Rails & Webpack concurrently
-rails:   rails s -e development -p ${API_PORT:-3000} -b 0.0.0.0
-webpack: ./node_modules/.bin/webpack-dev-server --config config/webpack.config.js
-worker1: rake jobs:work
-worker2: rake jobs:work
-logger:  rails r lib/log_service.rb
+rails:            rails s -e development -p ${API_PORT:-3000} -b 0.0.0.0
+log_service:      rails r lib/log_service_runner.rb
+resource_service: rails r lib/resource_service_runner.rb
+webpack:          ./node_modules/.bin/webpack-dev-server --config config/webpack.config.js
+worker:           rake jobs:work
 
 # UNCOMMENT THIS LINE IF YOU ARE DOING MOBILE TESTING:
 # Get started with `npm install weinre -g`
