+++ lib/_log_service_support.rb
@@ -1,0 +1,16 @@
+require_relative "./log_service"
+
+begin
+  # Listen to all logs on the message broker and store them in the database.
+  Transport
+    .current
+    .log_channel
+    .subscribe(block: true) do |info, _, payl|
+      LogService.process(info, payl.force_encoding("UTF-8"))
+    end
+rescue StandardError => e
+  Rollbar.error(e)
+  puts "MQTT Broker is unreachable. Waiting 5 seconds..."
+  sleep 5
+  retry
+end
