+++ lib/resource_service_runner.rb
@@ -1,0 +1,6 @@
+require_relative "./resource_service"
+
+loop do
+  sleep 1
+  puts "This is a noop."
+end
